requirements.txt
censys==2.2.16
python-dotenv==1.0.1
rich==13.9.4
threatfox==1.0.0
ipinfo==5.1.1
matplotlib==3.10
requests==2.32.3
PyQt5==5.15.10